Added Buttons
Added buttons and labels for different colored lines. No Ids or classes yet, but it is ready to be assigned stuff as we design the logic for it.

I also changed the header color to make the site a bit more readable and made the map to fit more of the screen.